+++ ananse/influence.py
@@ -1,4 +1,6 @@
 """Predict TF influence score"""
+import os
+import shutil
 import sys
 import warnings
 import genomepy
@@ -13,6 +15,7 @@ from sklearn.preprocessing import minmax_scale
 from scipy.stats import rankdata, mannwhitneyu
 
 from . import SEPARATOR
+from .utils import mytmpdir
 
 
 warnings.filterwarnings("ignore")
@@ -21,26 +24,40 @@ warnings.filterwarnings("ignore")
 Expression = namedtuple("Expression", ["score", "absfc", "realfc"])
 
 
-def read_top_interactions(fname, edges=100000, GRNsort_column="prob"):
-    """Read network file and return the top interactions
-    by default it incorporates all the data (prob score) to select the most
-    certain interactions, however other sorting options such as the weighted binding score"""
+def read_top_interactions(fname, edges=100_000, sort_by="prob"):
+    """
+    Read network file and return the top interactions.
+
+    Incorporates the combined metric (probability score) by default to select the most
+    certain interactions, however other sorting options such as the weighted binding score
+    can be applied too.
+    """
+    # read the GRN file
     rnet = pd.read_csv(
-        fname, usecols=["tf_target", GRNsort_column], sep="\t"
-    )  # read the GRN file
-    rnet.sort_values(
-        GRNsort_column, ascending=False, inplace=True
-    )  # sort based on the selected value (default probability) score
+        fname,
+        usecols=["tf_target", sort_by],
+        sep="\t",
+        dtype={"tf_target": str, sort_by: "float32"},
+    )
+    # sort based on the selected value (default probability score)
+    rnet.sort_values(sort_by, ascending=False, inplace=True, ignore_index=True)
+    # takes all edges if there are less than requested
     rnet = rnet.head(edges)
-    top_int = set(rnet["tf_target"])
+    top_int = set(rnet["tf_target"].unique())  # adding unique() speeds it up!
     return top_int
 
 
 def read_network(
-    fname, edges=100000, interactions=None, GRNsort_column="prob", full_output=False
+    fname, edges=100_000, interactions=None, sort_by="prob", full_output=False
 ):
-    """Read network file and return networkx DiGraph"""
+    """
+    Read a network file and return a networkx DiGraph.
 
+    Subsets the graph to interactions if given, else to the number top interactions given by edges.
+    If both interactions and edges are none, return the whole graph.
+
+    Peak memory usage is about ~5GB per million edges (tested with 0.1m, 1m and 10m edges)
+    """
     # read GRN files
     data_columns = ["tf_target", "prob"]
     if full_output:
@@ -52,172 +69,227 @@ def read_network(
             "weighted_binding",
             "activity",
         ]
+    # read the GRN file
     rnet = pd.read_csv(
         fname,
         sep="\t",
         usecols=data_columns,
         dtype="float64",
         converters={"tf_target": str},
-    )  # read the GRN file
+    )
     # sort on selection variable
-    rnet.sort_values(GRNsort_column, ascending=False, inplace=True)
-    if interactions is None:
-        rnet = rnet.head(edges)  # no interactions so take the top head of interactions
-    else:
+    rnet.sort_values(sort_by, ascending=False, inplace=True)
+    if interactions is not None:
         rnet = rnet[rnet.tf_target.isin(interactions)]
+    elif edges is not None:
+        rnet = rnet.head(edges)
 
-    G = nx.DiGraph()  # initiate empty network
-    for _, row in rnet.iterrows():
-        source, target = row[0].split(SEPARATOR, 1)
-        try:
-            if full_output:
-                G.add_edge(
-                    source,
-                    target,
-                    weight=row["prob"],
-                    weighted_binding=row["weighted_binding"],
-                    tf_expression=row["tf_expression"],
-                    tg_expression=row["target_expression"],
-                    tf_activity=row["activity"],
-                )
-            else:
-                G.add_edge(source, target, weight=row["prob"], n=1)
-        except Exception:
-            logger.error(f"Could not parse edge weight {source}:{target}")
-            raise
-    return G
+    # split the transcription factor and target gene into 2 columns
+    rnet[["source", "target"]] = rnet["tf_target"].str.split(SEPARATOR, expand=True)
+    rnet.drop(["tf_target"], inplace=True, axis=1)
+
+    # rename the columns
+    rnet.rename(
+        columns={
+            "prob": "weight",
+            # "weighted_binding": "weighted_binding",
+            # "tf_expression": "tf_expression",
+            "target_expression": "tg_expression",
+            "activity": "tf_activity",
+        },
+        inplace=True,
+    )
+
+    # load into a network with TFs and TGs as nodes, and the interaction scores as edges
+    grn = nx.from_pandas_edgelist(rnet, edge_attr=True, create_using=nx.DiGraph)
+
+    return grn
 
 
-def difference(GRN_source, GRN_target, full_output=False):
-    """Calculate the network different between two GRNs.
-    It first takes the nodes from both networks, and subsequently
-    first adds the edges from the target network that are missing in
-    the source network.
-    Secondly it add the edges present in both but with a higher interaction
-    score in the target network
+def difference(grn_source, grn_target, full_output=False):
     """
-    DIF = nx.create_empty_copy(GRN_source)  # copying source GRN nodes
-    nodes_target = nx.create_empty_copy(GRN_target)  # copying target GRN nodes
-    # merge the nodes if there are differences
-    # (which can happen) when taking the header instead of the --union-interactions flag
-    DIF.add_nodes_from(list(nodes_target.nodes))
+    Calculate the network different between two GRNs.
 
-    # lets check if the full GRN output is loaded and if so output all attributes to the diffnetwork:
+    First take the nodes from both networks, and add
+    edges from the target network that are missing in the source network.
+    Then add edges present in both but with a higher interaction
+    score in the target network.
+    """
+    grn_diff = nx.DiGraph()
     if full_output:
         logger.info("Calculating differential GRN with full output")
-        # lets load all the  edges into the diffnetwork
-        for u, v, ddict in GRN_target.edges(data=True):
+        # load all the edges into the diffnetwork
+        for u, v, ddict in grn_target.edges(data=True):
             # u = source node, v = target node, ddict = dictionary of edge attributes
             # calculate the weight difference and output all attributes
-            source_weight = GRN_source.edges[u, v]["weight"]
-            target_weight = GRN_target.edges[u, v]["weight"]
-            diff_weight = target_weight - source_weight
-            if diff_weight > 0:
+            weight_source = grn_source.edges[u, v]["weight"]
+            weight_target = grn_target.edges[u, v]["weight"]
+            weight_diff = weight_target - weight_source
+            if weight_diff > 0:
                 # if the interaction probability is higher in the target than in the
                 # source, add the interaction to the diffnetwork:
-                DIF.add_edge(
+                grn_diff.add_edge(
                     u,
                     v,
-                    weight=diff_weight,
-                    source_weight=source_weight,
-                    target_weight=target_weight,
-                    neglogweight=-np.log(diff_weight),
-                    n=1,
+                    weight=weight_diff,
+                    weight_source=weight_source,
+                    weight_target=weight_target,
                     tf_expr_diff=(
-                        (ddict["tf_expression"]) - (GRN_source[u][v]["tf_expression"])
+                        (ddict["tf_expression"]) - (grn_source[u][v]["tf_expression"])
                     ),
+                    tf_expr_source=grn_source[u][v]["tf_expression"],
                     tf_expr_target=ddict["tf_expression"],
-                    tf_expr_source=GRN_source[u][v]["tf_expression"],
                     tg_expr_diff=(
-                        (ddict["tg_expression"]) - (GRN_source[u][v]["tg_expression"])
+                        (ddict["tg_expression"]) - (grn_source[u][v]["tg_expression"])
                     ),
+                    tg_expr_source=grn_source[u][v]["tg_expression"],
                     tg_expr_target=ddict["tg_expression"],
-                    tg_expr_source=GRN_source[u][v]["tg_expression"],
                     wb_diff=(
                         (ddict["weighted_binding"])
-                        - (GRN_source[u][v]["weighted_binding"])
+                        - (grn_source[u][v]["weighted_binding"])
                     ),
-                    target_wb=ddict["weighted_binding"],
-                    source_wb=GRN_source[u][v]["weighted_binding"],
-                    TF_act_diff=(
-                        (ddict["tf_activity"]) - (GRN_source[u][v]["tf_activity"])
+                    wb_source=grn_source[u][v]["weighted_binding"],
+                    wb_target=ddict["weighted_binding"],
+                    tf_act_diff=(
+                        (ddict["tf_activity"]) - (grn_source[u][v]["tf_activity"])
                     ),
-                    TF_act_target=ddict["tf_activity"],
-                    TF_act_source=GRN_source[u][v]["tf_activity"],
+                    tf_act_source=grn_source[u][v]["tf_activity"],
+                    tf_act_target=ddict["tf_activity"],
                 )
     else:
         logger.info("Calculating differential GRN")
         # if only the weight is loaded, lets load only that in the diffnetwork:
-        for (u, v) in GRN_target.edges():
-            source_weight = GRN_source.edges[u, v]["weight"]
-            target_weight = GRN_target.edges[u, v]["weight"]
-            diff_weight = target_weight - source_weight
-            if diff_weight > 0:
-                DIF.add_edge(
-                    u, v, weight=diff_weight, n=1, neglogweight=-np.log(diff_weight)
+        for (u, v) in grn_target.edges():
+            weight_source = grn_source.edges[u, v]["weight"]
+            weight_target = grn_target.edges[u, v]["weight"]
+            weight_diff = weight_target - weight_source
+            if weight_diff > 0:
+                grn_diff.add_edge(
+                    u,
+                    v,
+                    weight=weight_diff,
                 )
-    return DIF
+    return grn_diff
 
 
-def targetScore(node, G, expression_change, max_degree=3):
-    """Calculate the influence score."""
-    total_score = 0
-    # Get the targets that are within a certain number of steps from TF
-    lengths, paths = nx.single_source_dijkstra(G, node, cutoff=max_degree - 1)
-    targets = [t for t in lengths if 0 < lengths[t] <= max_degree]
-    for target in paths:
+def get_weight(grn, path):
+    weight = np.cumprod([grn[s][t]["weight"] for s, t in zip(path, path[1:])])[-1]
+    # Add weight correction for the length of the path
+    weight = weight / (len(path) - 1)
+    return weight
+
+
+def target_score(node, grn, expression_change, targets):
+    """
+    Calculate the target score, as (mostly) explained in equation 5:
+    https://academic.oup.com/nar/article/49/14/7966/6318498#M5
+    """
+    ts = 0
+    for target in targets:
+        # Calculate all paths from TF to target to select to path with the highest multiplied weight
         all_paths = {}
-        # Calculate all paths from TF to target to select to path with the lowest total weight
-        for path in nx.all_simple_paths(G, node, target, cutoff=max_degree - 1):
-            if len(path) <= max_degree:
-                weight = np.cumprod(
-                    [G[s][t]["weight"] for s, t in zip(path, path[1:])]
-                )[-1]
-                # Add weight, corrected for the length of the path
-                all_paths[tuple(path)] = weight / (len(path) - 1)
+        for path in nx.all_simple_paths(grn, node, target, cutoff=2):
+            all_paths[tuple(path)] = get_weight(grn, path)
         if len(all_paths) > 0:
-            path, weight = sorted(all_paths.items(), key=lambda p: p[1])[-1]
+            path, weight = sorted(all_paths.items(), key=lambda pw: pw[1])[-1]
             # the level (or the number of steps) that gene is away from transcription factor
+            # TODO: this is the number of nodes. 1 higher than the number of steps!
             pathlen = len(path)
             # expression score of the target
             g = expression_change[target].score if target in expression_change else 0
+            # TODO: why divide by path length AGAIN? (already happens in the weight function)
+            # TODO: maybe this was left in by accident?
+            # TODO: see https://github.com/vanheeringen-lab/ANANSE/commit/ba67ebb8e7bafd1df13fb439485b6a590482e924
             score = g / pathlen * weight
-            total_score += score
+            ts += score
+    return ts
 
-    # Get Mann-Whitney U p-value of direct targets vs. non-direct targets
-    direct_targets = [n for n in G[node] if n in expression_change]
-    non_direct_targets = [
-        n for n in list(G.nodes) if n in expression_change and n not in direct_targets
-    ]
 
-    target_fc = [expression_change[t].absfc for t in direct_targets]
-    non_target_fc = [expression_change[t].absfc for t in non_direct_targets]
+# TODO: this version is ~O(n^2) faster, and returns near identical scores
+# def target_score(grn, expression_change, targets):
+#     ts = 0
+#     for target in targets:
+#         path = targets[target]
+#         weight = get_weight(grn, path)
+#         score = expression_change[target].score / len(path) * weight
+#         ts += score
+#     return ts
 
-    try:
-        pval = mannwhitneyu(target_fc, non_target_fc)[1]
-    except (RecursionError, ValueError):
-        pval = np.NAN
-        logger.warning(
-            f"Could not calculate p-val (target vs non-target fold-change) for {node}, "
-            f"targets = {len(target_fc)}, non-target = {len(non_target_fc)}."
-        )
-        logger.warning(f"targets = {target_fc[0:min(5, len(target_fc))]}...")
-        logger.warning(f"non_target = {non_target_fc[0:min(5, len(non_target_fc))]}...")
-    target_fc_diff = np.mean(target_fc) - np.mean(non_target_fc)
 
-    # factor, targetScore, directTargets, totalTargets, Gscore, pval, target_fc
+def influence_scores(node, grn, expression_change, de_genes):
+    """
+    Calculate the influence scores of a transcription factor.
+
+    Parameters
+    ----------
+    node : str
+        Transcription factor name, present in grn as a node
+    grn : nx.DiGraph
+        A network with gene names as nodes and interaction scores as weights
+    expression_change : dict
+        A dictionary with interaction scores and log fold changes per transcription factor
+
+    Returns
+    -------
+    tuple
+        interaction data of the given transcription factor
+    """
+    # get all genes that are
+    # - direct or indirectly targeted by the TF (cutoff + weight)
+    # - not the TF itself (because we divide by len(path)-1 elsewhere)
+    # - differentially expressed
+    targets = nx.single_source_dijkstra(grn, node, cutoff=2, weight=None)[1]  # noqa
+    _ = targets.pop(node, None)
+
+    de_targets = {k: v for k, v in targets.items() if k in de_genes}
+    targetscore = target_score(node, grn, expression_change, de_targets)
+
+    pval, target_fc_diff = fold_change_scores(node, grn, expression_change)
+    factor_fc = expression_change[node].absfc if node in expression_change else 0
     return (
-        node,
-        total_score,
-        G.out_degree(node),
-        len(targets),
-        expression_change[node].absfc if node in expression_change else 0,
-        pval,
-        target_fc_diff,
+        node,  # factor
+        grn.out_degree(node),  # noqa. directTargets
+        len(targets),  # totalTargets
+        targetscore,  # targetscore
+        expression_change[node].score,  # Gscore
+        factor_fc,  # factor_fc
+        pval,  # pval
+        target_fc_diff,  # target_fc
     )
 
 
-def filter_TF(scores_df, network=None, tpmfile=None, tpm=20, overlap=0.98):
+def fold_change_scores(node, grn, expression_change):
+    """
+    Get the Mann-Whitney U p-value of direct targets vs. non-direct targets,
+    as well as the difference of the mean fold changes.
+    """
+    direct_targets = set(grn[node]) & set(expression_change)
+    if len(direct_targets) == 0:
+        return np.NAN, np.NAN
+    non_direct_targets = (set(grn.nodes) & set(expression_change)) - direct_targets
+    if len(non_direct_targets) == 0:
+        return np.NAN, np.NAN
+
+    target_fc = [expression_change[t].absfc for t in direct_targets]
+    non_target_fc = [expression_change[t].absfc for t in non_direct_targets]
+    try:
+        # asymptotic method prevents recursion errors.
+        # TODO: review when scipy closes https://github.com/scipy/scipy/issues/14622
+        pval = mannwhitneyu(target_fc, non_target_fc, method="asymptotic")[1]
+    except (RecursionError, ValueError) as e:
+        pval = np.NAN
+        logger.warning(e)
+        # logger.warning(
+        #     f"Could not calculate p-val (target vs non-target fold-change) for {node}, "
+        #     f"targets = {len(target_fc)}, non-target = {len(non_target_fc)}."
+        # )
+        # logger.warning(f"targets = {target_fc[0:min(5, len(target_fc))]}...")
+        # logger.warning(f"non_target = {non_target_fc[0:min(5, len(non_target_fc))]}...")
+    target_fc_diff = np.mean(target_fc) - np.mean(non_target_fc)
+    return pval, target_fc_diff
+
+
+def filter_tf(scores_df, network=None, tpmfile=None, tpm=20, overlap=0.98):
     """Filter TFs:
     1) it have high expression in origin cell type;
     2) 98% of its target genes are also regulated by previous TFs.
@@ -257,61 +329,60 @@ class Influence(object):
         outfile,
         degenes,
         gene_gtf=None,
-        GRN_source_file=None,
-        GRN_target_file=None,
-        filter=False,  # TODO: variable not exposed in CLI
+        grn_source_file=None,
+        grn_target_file=None,
+        filter_tfs=False,  # TODO: variable not exposed in CLI
         edges=100_000,
         ncore=1,
-        GRNsort_column="prob",
+        sort_by="prob",
         padj_cutoff=0.05,
         full_output=False,
     ):
         self.ncore = ncore
         self.gene_gtf = gene_gtf
         self.full_output = full_output
-        self.GRNsort_column = GRNsort_column
+        self.outfile = outfile
+        self.filter_tfs = filter_tfs
 
         # Load GRNs
-        if GRN_target_file is None and GRN_source_file is None:
+        if grn_target_file is None and grn_source_file is None:
             logger.error("You should provide at least one ANANSE network file!")
             sys.exit(1)
         logger.info(f"Loading network data, using the top {edges} edges")
-        if GRN_source_file is None:
-            self.G = read_network(GRN_target_file, edges=edges)
+        if grn_source_file is None:
+            self.grn = read_network(grn_target_file, edges=edges)
             logger.warning("You only provided the target network!")
-        elif GRN_target_file is None:
-            self.G = read_network(GRN_source_file, edges=edges)
+        elif grn_target_file is None:
+            self.grn = read_network(grn_source_file, edges=edges)
             logger.warning("You only provided the source network!")
         else:
             top_int_source = read_top_interactions(
-                GRN_source_file, edges=edges, GRNsort_column=GRNsort_column
+                grn_source_file, edges=edges, sort_by=sort_by
             )
             top_int_target = read_top_interactions(
-                GRN_target_file, edges=edges, GRNsort_column=GRNsort_column
+                grn_target_file, edges=edges, sort_by=sort_by
             )
+            # combine the top edges of each input network
             top_int = set.union(top_int_source, top_int_target)
-            G1_source = read_network(
-                GRN_source_file,
+            grn_source = read_network(
+                grn_source_file,
                 interactions=top_int,
-                GRNsort_column=GRNsort_column,
+                sort_by=sort_by,
                 full_output=full_output,
             )
-            G2_target = read_network(
-                GRN_target_file,
+            grn_target = read_network(
+                grn_target_file,
                 interactions=top_int,
-                GRNsort_column=GRNsort_column,
+                sort_by=sort_by,
                 full_output=full_output,
             )
-            self.G = difference(G1_source, G2_target, full_output=full_output)
-            if len(self.G.edges) == 0:
+            self.grn = difference(grn_source, grn_target, full_output=full_output)
+            if len(self.grn.edges) == 0:
                 raise ValueError("No differences between networks!")
-            logger.info(f"    Differential network has {len(self.G.edges)} edges.")
+            logger.info(f"    Differential network has {len(self.grn.edges)} edges.")
 
         # Load expression file
         self.expression_change = self.read_expression(degenes, padj_cutoff)
-        self.outfile = outfile
-        # Filter TFs
-        self.filter = filter
 
     def read_expression(self, fname, padj_cutoff=0.05):
         """
@@ -355,7 +426,7 @@ class Influence(object):
         df = df[["log2FoldChange", "padj"]].dropna()  # removes unneeded data
         df = df.astype(float)
 
-        network_genes = set(self.G.nodes)
+        network_genes = set(self.grn.nodes)
         pct_overlap = len(network_genes & set(df.index)) / len(network_genes)
         logger.debug(
             f"{int(100 * pct_overlap)}% of genes found in DE genes and network(s)"
@@ -430,11 +501,11 @@ class Influence(object):
     def save_reg_network(self, filename, full_output=False):
         """Save the network difference between two cell types to a file."""
         # check if all keys are found
-        n = list(self.G.edges)[0]
-        keys = self.G.edges[n].keys()
+        n = list(self.grn.edges)[0]
+        keys = self.grn.edges[n].keys()
         with open(filename, "w") as nw:
-            if full_output and "source_wb" in keys:
-                logger.info("output full diff network")
+            if full_output and "wb_source" in keys:
+                logger.info("    Writing full differential network.")
                 header = [
                     "tf",
                     "target",
@@ -451,17 +522,17 @@ class Influence(object):
                     "wb_source",
                     "wb_target",
                     "tf_act_diff",
-                    "source_tf_act",
-                    "target_tf_act",
+                    "tf_act_source",
+                    "tf_act_target",
                 ]
                 nw.write("\t".join(header) + "\n")
-                for (u, v, ddict) in self.G.edges(data=True):
+                for (u, v, ddict) in self.grn.edges(data=True):
                     cols = [
                         u,
                         v,
                         ddict["weight"],
-                        ddict["source_weight"],
-                        ddict["target_weight"],
+                        ddict["weight_source"],
+                        ddict["weight_target"],
                         ddict["tf_expr_diff"],
                         ddict["tf_expr_source"],
                         ddict["tf_expr_target"],
@@ -469,117 +540,102 @@ class Influence(object):
                         ddict["tg_expr_source"],
                         ddict["tg_expr_target"],
                         ddict["wb_diff"],
-                        ddict["source_wb"],
-                        ddict["target_wb"],
-                        ddict["TF_act_diff"],
-                        ddict["TF_act_source"],
-                        ddict["TF_act_target"],
+                        ddict["wb_source"],
+                        ddict["wb_target"],
+                        ddict["tf_act_diff"],
+                        ddict["tf_act_source"],
+                        ddict["tf_act_target"],
                     ]
                     nw.write("\t".join(str(v) for v in cols) + "\n")
             else:
                 nw.write("tf\ttarget\tweight\n")
-                for (u, v, ddict) in self.G.edges(data=True):
+                for (u, v, ddict) in self.grn.edges(data=True):
                     nw.write(u + "\t" + v + "\t" + str(ddict["weight"]) + "\n")
 
-    def run_target_score(self, max_degree=3):
+    def run_target_score(self):
         """Run target score for all TFs."""
 
-        tfs = [node for node in self.G.nodes() if self.G.out_degree(node) > 0]
+        tfs = [
+            node for node in self.grn.nodes() if self.grn.out_degree(node) > 0  # noqa
+        ]
         logger.info(f"Differential network contains {len(tfs)} transcription factors.")
 
         # differentially expressed TFs
-        detfs = [tf for tf in tfs if tf in self.expression_change]
-
-        if len(detfs) == 0:
+        de_tfs = [tf for tf in tfs if tf in self.expression_change]
+        if len(de_tfs) == 0:
             logger.error(
                 "No overlapping transcription factors found between the network file(s) "
                 "(-s/--source, -t/--target) and the differential expression data (-d/--degenes)!"
             )
             sys.exit(1)
 
-        detfs = [tf for tf in detfs if self.expression_change[tf].realfc > 0]
-        if len(detfs) == 0:
-            logger.error(
-                "No differentially expressed TFs found with a log2 fold change above 0!"
-            )
+        # TODO: should 'realfc' this not be 'score' (padj<cutoff), or even 'absfc'?
+        de_tfs = set(tf for tf in de_tfs if self.expression_change[tf].realfc > 0)
+        if len(de_tfs) == 0:
+            # expression_change[tf].score > 0 == differentially expressed
+            logger.error("No increasingly expressed TFs found!")
             sys.exit(1)
         else:
-            logger.info(f"Out of these, {len(detfs)} are differentially expressed.")
+            logger.info(f"    Out of these, {len(de_tfs)} are increasingly expressed.")
 
-        result = []
-        if self.ncore > 1:
-            try:
-                pool = mp.Pool(self.ncore)
-                jobs = []
-                for tf in detfs:
-                    jobs.append(
-                        pool.apply_async(
-                            targetScore,
-                            (tf, self.G, self.expression_change, max_degree),
-                        )
-                    )
-                with tqdm(total=len(jobs)) as pbar:
-                    for j in jobs:
-                        result.append(j.get())
-                        pbar.update(1)
-                pool.close()
-            except Exception as e:
-                if "multiprocessing" in e.__repr__():
-                    logger.error(str(e))
-                    logger.error("Error seems to be related to multiprocessing.")
-                    logger.error(
-                        "In some cases running `ananse influence` with `-n 1` will solve this issue."
-                    )
-                    logger.error(
-                        "If it doesn't, then please file a bug report (with the output of the command run with `-n 1`) at:"
-                    )
-                    logger.error("https://github.com/vanheeringen-lab/ANANSE/issues")
-                    sys.exit(1)
-                raise
-        else:
-            for tf in tqdm(detfs):
-                result.append(
-                    targetScore(tf, self.G, self.expression_change, max_degree)
-                )
+        # differentially expressed genes
+        genes = self.grn.nodes
+        logger.info(f"Differential network contains {len(genes)} genes.")
+        de_genes = set(
+            g
+            for g in self.expression_change
+            if g in genes and self.expression_change[g].score > 0
+        )
+        logger.info(f"    Out of these, {len(de_genes)} are differentially expressed.")
 
-        # Get results and write to file
-        influence_file = open(self.outfile, "w")
+        tmpdir = mytmpdir()
+        tmpfile = os.path.join(tmpdir, os.path.basename(self.outfile))
+        influence_file = open(tmpfile, "w")
         influence_file.write(
             "factor\tdirectTargets\ttotalTargets\ttargetscore\tGscore\tfactor_fc\tpval\ttarget_fc\n"
         )
 
-        for (
-            factor,
-            score,
-            direct_targets,
-            total_targets,
-            factor_fc,
-            pval,
-            target_fc,
-        ) in result:
-            print(
-                factor,
-                direct_targets,
-                total_targets,
-                score,
-                self.expression_change[factor].score,
-                factor_fc,
-                pval,
-                target_fc,
-                file=influence_file,
-                sep="\t",
-            )
-        print("\n", file=influence_file)
+        try:
+            if self.ncore > 1:
+                pool = mp.Pool(self.ncore)
+                jobs = []
+                for tf in de_tfs:
+                    jobs.append(
+                        pool.apply_async(
+                            influence_scores,
+                            (tf, self.grn, self.expression_change, de_genes),
+                        )
+                    )
+                pool.close()
+                with tqdm(total=len(jobs)) as pbar:
+                    for j in jobs:
+                        print(*j.get(), file=influence_file, sep="\t")
+                        pbar.update(1)
+                pool.join()
 
-        influence_file.close()
+            else:
+                for tf in tqdm(de_tfs):
+                    line = influence_scores(
+                        tf, self.grn, self.expression_change, de_genes
+                    )
+                    print(*line, file=influence_file, sep="\t")
 
-        scores_df = pd.read_table(self.outfile, index_col=0)
-        scores_df["targetScaled"] = minmax_scale(
-            rankdata(scores_df["targetscore"], method="dense")
-        )
-        scores_df.sort_values("targetScaled", inplace=True, ascending=False)
+            influence_file.close()
+            shutil.move(tmpfile, self.outfile)
 
-        return self.outfile
+        except Exception as e:
+            pool = None  # noqa: force garbage collection on orphaned workers
+            if "multiprocessing" in e.__repr__():
+                msgs = [
+                    str(e),
+                    "The error seems to be related to multiprocessing.",
+                    "In some cases running `ananse influence` with `-n 1` will solve this issue.",
+                    "If it doesn't, please file a bug report (with the output of the command run with `-n 1`) at:",
+                    "https://github.com/vanheeringen-lab/ANANSE/issues",
+                ]
+                _ = [logger.error(msg) for msg in msgs]
+                sys.exit(1)
+            raise e
 
     def run_influence_score(self, influence_file, fin_expression=None):
         """Calculate influence score from target score and gscore"""
@@ -612,9 +668,9 @@ class Influence(object):
 
         scores_df.to_csv(self.outfile, sep="\t")
 
-        if self.filter:
-            scores_df2 = filter_TF(
-                network=self.G, scores_df=scores_df, tpmfile=fin_expression
+        if self.filter_tfs:
+            scores_df2 = filter_tf(
+                network=self.grn, scores_df=scores_df, tpmfile=fin_expression
             )
             scores_df2.to_csv(
                 ".".join(self.outfile.split(".")[:-1]) + "_filtered.txt", sep="\t"
@@ -622,13 +678,11 @@ class Influence(object):
 
     def run_influence(self, fin_expression=None):
         logger.info("Saving differential network.")
-        self.save_reg_network(
-            (".".join(self.outfile.split(".")[:-1]) + "_diffnetwork.tsv"),
-            full_output=self.full_output,
-        )
+        diffnetwork = os.path.splitext(self.outfile)[0] + "_diffnetwork.tsv"
+        self.save_reg_network(diffnetwork, self.full_output)
 
         logger.info("Calculating target scores.")
-        influence_file = self.run_target_score()
+        self.run_target_score()
 
         logger.info("Calculating influence scores.")
-        self.run_influence_score(influence_file, fin_expression=fin_expression)
+        self.run_influence_score(self.outfile, fin_expression)